pyproject.toml
[tool.poetry]
name = "acronymbot"
version = "0.1.0"
description = ""
authors = ["Jessica Walkenhorst"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.4.1"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.9.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
